# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 관리의달인 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfqs:
    get:
      operationId: listRfqs
      tags: [vendors]
      summary: List RFQs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, cancelled]
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: forVendorId
          in: query
          description: Filter RFQs visible to this vendor (by invitation + geo/category match)
          schema:
            type: integer
      responses:
        "200":
          description: RFQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rfq"
    post:
      operationId: createRfq
      tags: [vendors]
      summary: Create an RFQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRfqBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /rfqs/{id}:
    get:
      operationId: getRfq
      tags: [vendors]
      summary: Get a single RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RFQ detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
        "404":
          description: Not found
    patch:
      operationId: updateRfq
      tags: [vendors]
      summary: Update an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRfqBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
        "404":
          description: Not found
    delete:
      operationId: deleteRfq
      tags: [vendors]
      summary: Delete an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /rfqs/{id}/expand-scope:
    patch:
      operationId: expandRfqScope
      tags: [vendors]
      summary: Expand RFQ geo scope from sigungu to sido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expanded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
        "404":
          description: Not found
  /rfqs/{id}/matched-vendors:
    get:
      operationId: getRfqMatchedVendors
      tags: [vendors]
      summary: Get geo-matched vendors for an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matched vendor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"

  # ── Alert Actions (알림 처리) ─────────────────────────
...
  /work-reports:
    get:
      operationId: listWorkReports
      tags: [vendors]
      summary: List work reports
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [submitted, approved, rejected]
      responses:
        "200":
          description: Work report list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkReport"
    post:
      operationId: createWorkReport
      tags: [vendors]
      summary: Submit a work report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkReport"
  /work-reports/{id}:
    get:
      operationId: getWorkReport
      tags: [vendors]
      summary: Get a work report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkReport"
        "404":
          description: Not found
    patch:
      operationId: updateWorkReport
      tags: [vendors]
      summary: Review a work report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkReportBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkReport"
        "404":
          description: Not found

  # ── Settlements (정산) ──────────────────────────────────
...
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [elevator, water_tank, fire_safety, electrical, gas, septic, cleaning, security, waterproofing, maintenance_repair, defect_diagnosis, building_maintenance, mechanical, other]
        type:
          type: string
          enum: [contracted, platform]
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        isRecommended:
          type: boolean
        notes:
          type: ["string", "null"]
        businessRegNumber:
          type: ["string", "null"]
        representativeName:
          type: ["string", "null"]
        serviceArea:
          type: ["string", "null"]
        subCategories:
          type: ["string", "null"]
        sido:
          type: ["string", "null"]
        sigungu:
          type: ["string", "null"]
        joinedAt:
          type: ["string", "null"]
          format: date-time
        contractBuildingName:
          type: ["string", "null"]
        contractStartDate:
          type: ["string", "null"]
          format: date
        contractEndDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - type
        - isRecommended
        - createdAt
        - updatedAt
...
    Rfq:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [elevator, water_tank, fire_safety, electrical, gas, septic, cleaning, security, waterproofing, maintenance_repair, defect_diagnosis, building_maintenance, mechanical, other]
        description:
          type: ["string", "null"]
        buildingName:
          type: string
        desiredDate:
          type: ["string", "null"]
          format: date
        deadline:
          type: string
          format: date
        status:
          type: string
          enum: [open, closed, cancelled]
        vendorIds:
          type: ["string", "null"]
        sido:
          type: ["string", "null"]
        sigungu:
          type: ["string", "null"]
        geoScope:
          type: ["string", "null"]
        closeUpPhotoUrl:
          type: ["string", "null"]
        widePhotoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - buildingName
        - deadline
        - status
        - createdAt
        - updatedAt
    CreateRfqBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [elevator, water_tank, fire_safety, electrical, gas, septic, cleaning, security, waterproofing, maintenance_repair, defect_diagnosis, building_maintenance, mechanical, other]
        description:
          type: ["string", "null"]
        buildingName:
          type: string
        desiredDate:
          type: ["string", "null"]
          format: date
        deadline:
          type: string
          format: date
        vendorIds:
          type: ["string", "null"]
        sido:
          type: ["string", "null"]
        sigungu:
          type: ["string", "null"]
        geoScope:
          type: ["string", "null"]
        closeUpPhotoUrl:
          type: ["string", "null"]
        widePhotoUrl:
          type: ["string", "null"]
      required:
        - title
        - category
        - buildingName
        - deadline
    UpdateRfqBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [elevator, water_tank, fire_safety, electrical, gas, septic, cleaning, security, waterproofing, maintenance_repair, defect_diagnosis, building_maintenance, mechanical, other]
        description:
          type: ["string", "null"]
        buildingName:
          type: string
        desiredDate:
          type: ["string", "null"]
          format: date
        deadline:
          type: string
          format: date
        status:
          type: string
          enum: [open, closed, cancelled]
        vendorIds:
          type: ["string", "null"]
        sido:
          type: ["string", "null"]
        sigungu:
          type: ["string", "null"]
        geoScope:
          type: ["string", "null"]
...
    WorkReport:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        quoteId:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        completionDate:
          type: string
          format: date
        photoUrls:
          type: ["string", "null"]
        status:
          type: string
          enum: [submitted, approved, rejected]
        reviewNotes:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - rfqId
        - quoteId
        - vendorId
        - vendorName
        - title
        - completionDate
        - status
        - createdAt
        - updatedAt
    CreateWorkReportBody:
      type: object
      properties:
        rfqId:
          type: integer
        quoteId:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        completionDate:
          type: string
          format: date
        photoUrls:
          type: ["string", "null"]
      required:
        - rfqId
        - quoteId
        - vendorId
        - vendorName
        - title
        - completionDate
    UpdateWorkReportBody:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, approved, rejected]
        reviewNotes:
          type: ["string", "null"]
...